Added Coffee API mock

+++ coffee_app/lib/business/model/product.dart
@@ -20,7 +20,7 @@ class Product {
     name = json["name"];
     description = json["description"];
     imageUrl = json["imageUrl"];
-    price = json["price"];
+    price = json["price"].toDouble();
     categoryId = json["categoryId"];
   }
